Moved db rules to constants
