+++ mito-ai/mito_ai/completion_handlers/utils.py
@@ -1,32 +1,13 @@
 # Copyright (c) Saga Inc.
 # Distributed under the terms of the GNU Affero General Public License v3.0 License.
 
-import os
-import json
-from typing import Optional, Union, List, Dict, Any, cast, Final
+from typing import Optional, Union, List, Dict, Any, cast
 from mito_ai.message_history import GlobalMessageHistory
 from mito_ai.models import ThreadID
 from mito_ai.providers import OpenAIProvider
 from openai.types.chat import ChatCompletionMessageParam
 from mito_ai.prompt_builders.chat_system_message import create_chat_system_message_prompt
 from mito_ai.prompt_builders.agent_system_message import create_agent_system_message_prompt
-from mito_ai.utils.schema import MITO_FOLDER
-
-
-def get_database_configuration():
-    """
-    Returns any database configurations that the user has set up.
-    """
-
-    APP_DIR_PATH: Final[str] = os.path.join(MITO_FOLDER)
-    
-    with open(os.path.join(APP_DIR_PATH, 'db', 'connections.json'), 'r') as f:
-        connections = json.load(f)
-
-    with open(os.path.join(APP_DIR_PATH, 'db', 'schemas.json'), 'r') as f:
-        schemas = json.load(f)
-
-    return connections, schemas
 
 async def append_chat_system_message(
         message_history: GlobalMessageHistory,
@@ -38,8 +19,7 @@ async def append_chat_system_message(
     if any(msg["role"] == "system" for msg in message_history.get_ai_optimized_history(thread_id)):
         return
     
-    db_connections, db_schemas = get_database_configuration()
-    system_message_prompt = create_chat_system_message_prompt(db_connections, db_schemas)
+    system_message_prompt = create_chat_system_message_prompt()
 
     system_message: ChatCompletionMessageParam = {
         "role": "system",
