+++ mito-ai/mito_ai/prompt_builders/chat_system_message.py
@@ -1,9 +1,16 @@
 # Copyright (c) Saga Inc.
 # Distributed under the terms of the GNU Affero General Public License v3.0 License.
 
-from mito_ai.prompt_builders.prompt_constants import CITATION_RULES, ACTIVE_CELL_ID_SECTION_HEADING, CODE_SECTION_HEADING, FILES_SECTION_HEADING, VARIABLES_SECTION_HEADING
+from mito_ai.prompt_builders.prompt_constants import (
+    CITATION_RULES, 
+    ACTIVE_CELL_ID_SECTION_HEADING, 
+    CODE_SECTION_HEADING, 
+    FILES_SECTION_HEADING, 
+    VARIABLES_SECTION_HEADING, 
+    DATABASE_RULES
+)
 
-def create_chat_system_message_prompt(db_connections: dict, db_schemas: dict) -> str:
+def create_chat_system_message_prompt() -> str:
     return f"""You are Mito Data Copilot, an AI assistant for Jupyter. You're a great python programmer, a seasoned data scientist and a subject matter expert.
 
 The user is going to ask you for help writing code, debugging code, explaining code, or drawing conclusions from their data/graphs. It is your job to help them accomplish their goal. 
@@ -38,20 +45,7 @@ The average value is 53[MITO_CITATION:7b3a9e2c-5d14-4c83-b2f9-d67891e4a5f2:2]
 Notice in the example above that the citation uses line number 2 because citation line numbers are 0-indexed.
 
 ===
-DATABASE RULES:
-If the user has requested data that you belive is stored in the database:
-- Use the provided schema.
-- Only use SQLAlchemy to query the database.
-- Do not use a with statement when creating the SQLAlchemy engine. Instead, initialize it once so it can be reused for multiple queries.
-- Always return the results of the query in a pandas DataFrame, unless instructed otherwise.
-- Column names in query results may be returned in lowercase. Always refer to columns using their lowercase names in the resulting DataFrame (e.g., df['date'] instead of df['DATE']).
-- If you think the requested data is stored in the database, but you are unsure, then ask the user for clarification.
-
-Here is the schema:
-{db_schemas}
-
-Here are the connection details:
-{db_connections}
+{DATABASE_RULES}
 
 ==== 
 IMPORTANT RULES:
